Separated out all services into individual files to allow for applications to use portions of BACnet functionality, rather than needing to take everything.
Added additional bit level operations and added 64 bit counterparts to all current and new operations.

Moved the hash computation code out of the STL and into the BACnet library. Added 32 and 64 bit specific hashing code (hashing algorithm is the same, hash code is either 32 or 64 bits wide).

Added a hint file for interface macros so VS doesn't get as confused.

+++ BACnet/BACnetStringImpl.cpp
@@ -188,7 +188,7 @@ size_t ConvertToUTF32(const BACnetStringImpl * const s, U32 * const str, size_t 
 	// length == num codepoints == countof equivalent UTF-32 string.
 	// iterating over code unit "pairs" up to length will be safe unless we messed up the length calculation.
 	size_t index = 0;
-	for(int i = 0; i < slen; ++i)
+	for(size_t i = 0; i < slen; ++i)
 	{
 		str[i] = DecodeCodePoint16(&s->str[index]);
 		index += GetNumCodeUnits16(str[i]);
